Move admin lite related code into a separate app
