+++ src/core/settings/base.py
@@ -75,6 +75,7 @@ LOCAL_APPS = [
     'projects',
     'blog',
     'eligibility',
+    'admin_lite',
 ]
 
 INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS
